Standardise row comparison by making precision optional
The objective of the issue was to make allow_nan_equality available to
assert_approx_df_equality, but on reviewing the code it seemed like a
more efficient plan to include the precision option in
assert_df_equality and make it optional. This results in only one set
of functions needed to be maintained.

I've moved the single assert_rows_equality function into row_comparer
along with the custom exception which I've renamed. I've made one
generic equality function which takes options. All the tests have been
updated although there could be scope for a few more new tests - I
wanted to consult on that first.

This is a fairly large change to the API for which a major version
update would be recommended.